requirements.txt
# requirements.txt

pandas
numpy
scikit-learn
chardet
geopy
tabulate
openpyxl
pyyaml
jellyfish